usb: f_ipc: Add the flush_work when IPC disconnected
The QRTR will keep pushing the packet to USB until the USB transport is
unregistered from the QRTR layer. The USB de-registration from the QRTR
layer is happening in the work queue when there is a USB disconnect or
USB composition change occurs.

In this scenario, there is a possibility of memory list corruption if
QRTR continues to transfer the packets to USB even after the USB
connection is disconnected. This is because the QRTR layer is
unregistered in the work queue, which takes some time to execute.

Fix this issue by adding a flush work before returning the status from
USB to QRTR in USB disconnected scenario.

Change-Id: I436e40eaea389161cd6e889a7c636c6fda8749d5
Signed-off-by: Shashikala Katthi <quic_skatthi@quicinc.com>
(cherry picked from commit 881375a3092a032fb7acea4362df458ffb884e2a)