Revert "usb: f_ipc: Add the flush_work when IPC disconnected"
The commit <7f8cb9f9390cc> (“usb: f_ipc: Add the flush_work when IPC
disconnected”) caused a QTR IPC failure during the “adb root” test
scenario. The issue appears to be that unregister/register with QRTR
is not happening from USB during reconnect when using flush_work.
This has also led to watchdog failure on reboot test scenarios. The
reason for this is that flush_work called ipc read/write if USB got
disconnected. There is a chance of creating a blocking call if the work
is not scheduled after the trigger flush work in reboot scenario.

Change-Id: Iccd1233b37d5bed21ab657c852eec6aea29b51e7
Signed-off-by: Selvarasu Ganesan <quic_selvaras@quicinc.com>