cnss2: Do not stay wake if no event present in list
The pm_stay_awake() is called before processing events in cnss
driver event list without checking if the list is empty or not.
When the list is empty, the API pair pm_stay_awake/pm_relax is
called without doing anything. This will abort the on-going
suspend command which is not necessary.

Do not stay awake if there's no event to be processed in the event
list.

Change-Id: I4d2c99ebed04e9afccb60714929374e404d392b4
CRs-Fixed: 3620740
Signed-off-by: Guisen Yang <quic_guiseny@quicinc.com>